Refactor test folders for frost-core (#364)
Update ciphersuites due to refactoring in core
