+++ frost-ed448/src/tests/proptests.rs
@@ -1,9 +1,10 @@
 use frost_core::tests::proptests::{tweak_strategy, SignatureCase};
-use frost_ed25519::*;
 use proptest::prelude::*;
 
 use rand_chacha::ChaChaRng;
 use rand_core::SeedableRng;
+
+use crate::*;
 
 proptest! {
 
@@ -19,7 +20,7 @@ proptest! {
 
         // Create a test case for each signature type.
         let msg = b"test message for proptests";
-        let mut sig = SignatureCase::<Ed25519Sha512>::new(rng, msg.to_vec());
+        let mut sig = SignatureCase::<Ed448Shake256>::new(rng, msg.to_vec());
 
         // Apply tweaks to each case.
         for t in &tweaks {
