+++ frost-p256/src/keys/dkg.rs
@@ -95,6 +95,6 @@ mod tests {
     fn check_sign_with_dkg() {
         let rng = thread_rng();
 
-        frost_core::tests::check_sign_with_dkg::<P256Sha256, _>(rng);
+        frost_core::tests::ciphersuite_generic::check_sign_with_dkg::<P256Sha256, _>(rng);
     }
 }
