+++ frost-p256/src/tests.rs
@@ -5,6 +5,8 @@ mod coefficient_commitment_tests;
 #[cfg(test)]
 mod deserialize_tests;
 #[cfg(test)]
+mod proptests;
+#[cfg(test)]
 mod signing_tests;
 #[cfg(test)]
 mod vss_commitment_tests;
