+++ frost-p256/src/tests/vss_commitment_tests.rs
@@ -15,17 +15,19 @@ lazy_static! {
 #[test]
 fn check_serialize_vss_commitment() {
     let rng = thread_rng();
-    frost_core::tests::check_serialize_vss_commitment::<P256Sha256, _>(rng);
+    frost_core::tests::vss_commitment::check_serialize_vss_commitment::<P256Sha256, _>(rng);
 }
 
 #[test]
 fn check_deserialize_vss_commitment() {
     let rng = thread_rng();
-    frost_core::tests::check_deserialize_vss_commitment::<P256Sha256, _>(rng);
+    frost_core::tests::vss_commitment::check_deserialize_vss_commitment::<P256Sha256, _>(rng);
 }
 
 #[test]
 fn check_deserialize_vss_commitment_error() {
     let rng = thread_rng();
-    frost_core::tests::check_deserialize_vss_commitment_error::<P256Sha256, _>(rng, &ELEMENTS);
+    frost_core::tests::vss_commitment::check_deserialize_vss_commitment_error::<P256Sha256, _>(
+        rng, &ELEMENTS,
+    );
 }
